Convert full switcher to CLabels
